Add SPDX license list for checking against license field
This is the current complete SPDX license list, with code included for
regenerating it as needed in the future.

The purpose of this list is to be able to check that a SPDX license
name has been used, and if it is not valid, notify the user to
use a standardized name.

If the user has checked and there is no standardized identifier, then
they should give a URL to the source text of the license.